Move QrCodeGenerator to logic package
