+++ src/main/java/seedu/address/logic/QrCodeGenerator.java
@@ -1,4 +1,4 @@
-package seedu.address;
+package seedu.address.logic;
 
 import java.io.IOException;
 import java.nio.file.Files;
