+++ src/main/java/seedu/address/model/person/Person.java
@@ -12,9 +12,9 @@ import java.util.logging.Logger;
 
 import com.google.zxing.WriterException;
 
-import seedu.address.QrCodeGenerator;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.logic.QrCodeGenerator;
 import seedu.address.model.person.exceptions.AttributeNotFoundException;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.TagSet;
